+++ test/unit/test_memory.cpp
@@ -155,9 +155,9 @@ TEST(Memory, CharArrayToMemory) {
     auto canonical = MemoryToCanonicalString(memory);
     auto expected =
             "\
-00000000   54 68 65 20 71 75 69 63  6b 20 62 72 6f 77 6e 20   |The quic k brown |\n\
-00000010   66 6f 78 20 6a 75 6d 70  73 20 6f 76 65 72 20 74   |fox jump s over t|\n\
-00000020   68 65 20 6c 61 7a 79 20  64 6f 67                  |he lazy  dog     |\n";
+0x0000000000000000   54 68 65 20 71 75 69 63  6b 20 62 72 6f 77 6e 20   |The quic k brown |\n\
+0x0000000000000010   66 6f 78 20 6a 75 6d 70  73 20 6f 76 65 72 20 74   |fox jump s over t|\n\
+0x0000000000000020   68 65 20 6c 61 7a 79 20  64 6f 67                  |he lazy  dog     |\n";
 
     EXPECT_STREQ(canonical.c_str(), expected);
 }
@@ -171,9 +171,9 @@ TEST(Memory, StringToMemory) {
     auto canonical = MemoryToCanonicalString(memory);
     auto expected =
             "\
-00000000   54 68 65 20 71 75 69 63  6b 20 62 72 6f 77 6e 20   |The quic k brown |\n\
-00000010   66 6f 78 20 6a 75 6d 70  73 20 6f 76 65 72 20 74   |fox jump s over t|\n\
-00000020   68 65 20 6c 61 7a 79 20  64 6f 67                  |he lazy  dog     |\n";
+0x0000000000000000   54 68 65 20 71 75 69 63  6b 20 62 72 6f 77 6e 20   |The quic k brown |\n\
+0x0000000000000010   66 6f 78 20 6a 75 6d 70  73 20 6f 76 65 72 20 74   |fox jump s over t|\n\
+0x0000000000000020   68 65 20 6c 61 7a 79 20  64 6f 67                  |he lazy  dog     |\n";
 
     EXPECT_STREQ(canonical.c_str(), expected);
 }